+++ src/services/describe.ts
@@ -1,7 +1,7 @@
 import { storeRecordTypes, storePageLayoutItems, storeLocalization } from './salesforce/metadata';
 import { saveRecords, getAllRecords, getRecords, clearTable } from './database/database';
-
-import { RecordType, PageLayoutSection, PageLayoutItem, PicklistValue } from '../types/sqlite';
+import { getAllRecordTypes } from './database/metadata';
+import { SQLitePageLayoutSection, SQLitePageLayoutItem } from '../types/sqlite';
 import { SurveyLayout } from '../types/survey';
 
 import { logger } from '../utility/logger';
@@ -34,7 +34,6 @@ export const retrieveAllMetadata = async () => {
     const picklistValues = [...serializedPicklistValueSet.values()].map(s => {
       return JSON.parse(s as string);
     });
-    logger('DEBUG', 'picklistvalues refres', picklistValues);
     await saveRecords(DB_TABLE.PICKLIST_VALUE, picklistValues, undefined);
 
     // Field type object
@@ -47,22 +46,12 @@ export const retrieveAllMetadata = async () => {
         return result;
       }, {});
     storage.save({ key: ASYNC_STORAGE_KEYS.FIELD_TYPE, data: fieldType });
-
-    // Dictionary
+    // Localization
     await clearTable(DB_TABLE.Localization);
     await storeLocalization();
-    await buildDictionary();
-  } catch (e) {
-    console.log(e);
+  } catch {
+    throw new Error('Unexpected error occured while retrieving survey settings. Contact your administrator.');
   }
-};
-
-/**
- * @description Get all the record types of the survey object from local database
- */
-export const getAllRecordTypes = async (): Promise<Array<RecordType>> => {
-  const recordTypes: Array<RecordType> = await getAllRecords(DB_TABLE.RecordType);
-  return recordTypes;
 };
 
 /**
@@ -71,13 +60,13 @@ export const getAllRecordTypes = async (): Promise<Array<RecordType>> => {
  */
 export const buildLayoutDetail = async (layoutId: string): Promise<SurveyLayout> => {
   // sections in the layout
-  const sections: Array<PageLayoutSection> = await getRecords(
+  const sections: Array<SQLitePageLayoutSection> = await getRecords(
     DB_TABLE.PageLayoutSection,
     `where layoutId='${layoutId}'`
   );
   // items used in the sections
   const sectionIds = sections.map(s => s.id);
-  const items: Array<PageLayoutItem> = await getRecords(
+  const items: Array<SQLitePageLayoutItem> = await getRecords(
     DB_TABLE.PageLayoutItem,
     `where sectionId in (${sectionIds.map(id => `'${id}'`).join(',')})`
   );
@@ -105,59 +94,4 @@ export const buildLayoutDetail = async (layoutId: string): Promise<SurveyLayout>
   };
 
   return layout;
-};
-
-/**
- * @description
- * @param fieldName
- */
-export const getPicklistValues = async (fieldName: string) => {
-  const recordTypes: Array<PicklistValue> = await getRecords(DB_TABLE.PICKLIST_VALUE, `where fieldName='${fieldName}'`);
-  return recordTypes;
-};
-
-/**
- * @description Build expo-localization object from locally stored tables
- * @todo Remove duplicates in the table (i.e., field used in multiple layouts)
- */
-export const buildDictionary = async () => {
-  // original labels
-  const recordTypes = await getAllRecordTypes();
-  const recordTypeLabels = recordTypes.reduce((result, current) => {
-    result[`${L10N_PREFIX.RecordType}${current.name}`] = current.label;
-    return result;
-  }, {});
-  const sections = await getAllRecords(DB_TABLE.PageLayoutSection);
-  const sectionLabels = sections.reduce((result, current) => {
-    result[`${L10N_PREFIX.PageLayoutSection}${current.sectionLabel}`] = current.sectionLabel;
-    return result;
-  }, {});
-  const fields = await getAllRecords(DB_TABLE.PageLayoutItem);
-  const fieldLabels = fields.reduce((result, current) => {
-    result[`${L10N_PREFIX.PageLayoutItem}${current.fieldName}`] = current.fieldLabel;
-    return result;
-  }, {});
-
-  const originalLabels = { ...recordTypeLabels, ...sectionLabels, ...fieldLabels };
-  logger('FINE', 'buildDictionary', originalLabels);
-  // localization.
-  // TODO: create localization table first for no records in salesforce
-  const translatedRecordTypes = await getAllRecords(DB_TABLE.Localization);
-  const translatedLabels = translatedRecordTypes.reduce((result, current) => {
-    result[`${L10N_PREFIX[current.type]}${current.name}`] = current.label;
-    return result;
-  }, {});
-  logger('FINE', 'buildDictionary', translatedLabels);
-
-  i18n.translations = {
-    en: {
-      ...i18n.translations.en,
-      ...originalLabels,
-    },
-    ne: {
-      ...i18n.translations.ne,
-      ...translatedLabels,
-    },
-  };
-  // field
 };
